+++ spec/integrations/pro/consumption/strategies/aj/dlq_ftr_mom_vp/regular_processing_without_errors_spec.rb
@@ -45,7 +45,7 @@ consumer = setup_rdkafka_consumer
 consumer.subscribe(DT.topic)
 
 consumer.each do |message|
-  assert message.offset >= 30
+  assert message.offset >= 30, [DT, message.offset]
   break
 end