+++ spec/integrations/pro/consumption/strategies/vp/with_error_in_the_middle_marking_spec.rb
@@ -5,7 +5,7 @@
 
 class Listener
   def on_error_occurred(event)
-    DT[:errors] << event
+    DT[:errors] << true
   end
 end
 
@@ -40,7 +40,7 @@ draw_routes do
 end
 
 start_karafka_and_wait_until do
-  if DT[:offsets].size >= 50
+  if DT[:offsets].size >= 50 && DT[:offsets].count(25) >= 2
     true
   else
     produce_many(DT.topic, DT.uuids(10))